+++ modules/simple_forms_api/app/services/simple_forms_api/form_remediation/submission_remediation_data.rb
@@ -1,15 +1,13 @@
 # frozen_string_literal: true
 
-require 'simple_forms_api/form_submission_remediation/configuration/base'
+require 'simple_forms_api/form_remediation/configuration/base'
 
 module SimpleFormsApi
   module FormRemediation
     class SubmissionRemediationData
-      DEFAULT_CONFIG = SimpleFormsApi::FormSubmissionRemediation::Configuration::Base.freeze
-
       attr_reader :file_path, :submission, :attachments, :metadata
 
-      def initialize(id:, config: DEFAULT_CONFIG.new)
+      def initialize(id:, config: Configuration::Base.new)
         @config = config
 
         validate_input(id)
